% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_statistics.R
\name{analyze_statistics}
\alias{analyze_statistics}
\title{Perform group-wise t-test or ANOVA+TukeyHSD for qPCR}
\usage{
analyze_statistics(dct_long_sel, stat_method, short_labels, stat_dir)
}
\arguments{
\item{dct_long_sel}{Long-format data (delta_ct)}

\item{stat_method}{'ttest' or 'anova'}

\item{short_labels}{Sample names for comparison}

\item{stat_dir}{Output directory for csv results}
}
\value{
Data.frame (statistical results)
}
\description{
Perform group-wise t-test or ANOVA+TukeyHSD for qPCR
}
