% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ct_values.R
\name{calc_ct_values}
\alias{calc_ct_values}
\title{Calculate ΔCt, ΔΔCt, and relative expression values}
\usage{
calc_ct_values(wide, housekeeping, ref_sample, target_genes)
}
\arguments{
\item{wide}{Wide-format data.frame after outlier removal}

\item{housekeeping}{Name of housekeeping gene (string)}

\item{ref_sample}{Reference sample name (string)}

\item{target_genes}{Vector of target gene names}
}
\value{
Data.frame with added columns for delta_ct, delta_dct, rel_exp for each gene
}
\description{
Calculate ΔCt, ΔΔCt, and relative expression values
}
