% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_outliers}
\alias{check_outliers}
\title{Check and plot qPCR outliers}
\usage{
check_outliers(ct_data, meta_cols, outlier_dir, return_plot = FALSE)
}
\arguments{
\item{ct_data}{Data.frame from load_and_preprocess_rawdata}

\item{meta_cols}{Character vector of metadata columns}

\item{outlier_dir}{Output directory for result csv/plots}

\item{return_plot}{If TRUE, return ggplot object}
}
\value{
Data.frame with outlier flag or list (data + plot)
}
\description{
Check and plot qPCR outliers
}
