% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_preprocess_rawdata.R
\name{load_and_preprocess_rawdata}
\alias{load_and_preprocess_rawdata}
\title{Load and preprocess qPCR raw data}
\usage{
load_and_preprocess_rawdata(files, meta_cols)
}
\arguments{
\item{files}{Character vector of file paths (Excel).}

\item{meta_cols}{Character vector, names of metadata columns.}
}
\value{
A data.frame of tidy qPCR data.
}
\description{
Load and preprocess qPCR raw data
}
