% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_long_format.R
\name{to_long_format}
\alias{to_long_format}
\title{Convert wide-format qPCR data to long-format for plotting/statistics}
\usage{
to_long_format(wide_clean, meta_cols_full)
}
\arguments{
\item{wide_clean}{Output of calc_ct_values()}

\item{meta_cols_full}{Vector of meta-data column names to keep}
}
\value{
List with $rel_long and $dct_long (both long-format tibbles)
}
\description{
Convert wide-format qPCR data to long-format for plotting/statistics
}
