% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_comparison.R
\name{multi_comparison}
\alias{multi_comparison}
\title{Pipeline for multi-group qPCR comparison and plotting}
\usage{
multi_comparison(
  files,
  meta_cols,
  housekeeping,
  ref_sample,
  comparison_list,
  target_groups,
  stat_dir = "statistics",
  plot_dir = "plots",
  outlier_dir = "outlier",
  annotation_y_expand = 1.2,
  plot_height = 5,
  annotation_base = 0.08,
  annotation_step = 0.08
)
}
\arguments{
\item{files}{List of file paths to qPCR Excel files}

\item{meta_cols}{Metadata columns to include}

\item{housekeeping}{Housekeeping gene}

\item{ref_sample}{Reference sample}

\item{comparison_list}{List (per comparison) of list(name, selected_samples, short_labels, stat_method)}

\item{target_groups}{List of character vectors (target genes per comparison)}

\item{stat_dir}{Directory for stat csv files}

\item{plot_dir}{Directory for output plots}

\item{outlier_dir}{Directory for outlier results}

\item{...}{(additional params)}
}
\description{
Pipeline for multi-group qPCR comparison and plotting
}
